refactor: optimize imports for new directories
