+++ src/Donations/resources/components/DonationsListTable.tsx
@@ -1,10 +1,10 @@
 import {__, sprintf} from '@wordpress/i18n';
 import {useSWRConfig} from 'swr';
 import {ListTablePage} from '@givewp/components';
-import RowAction from '@givewp/components/ListTable/RowAction/RowAction';
+import RowAction from '@givewp/components/ListTable/RowAction';
 import ListTableApi from '@givewp/components/ListTable/api';
 import tableStyles from '@givewp/components/ListTable/ListTablePage/ListTablePage.module.scss';
-import {IdBadge} from '@givewp/components/ListTable/TableCell/TableCell';
+import {IdBadge} from '@givewp/components/ListTable/TableCell';
 import {BulkActionsConfig, FilterConfig, ShowConfirmModalContext} from '@givewp/components/ListTable/ListTablePage';
 import {useContext} from 'react';
 import {Interweave} from 'interweave';
