+++ src/Views/Components/ListTable/ListTablePage/index.tsx
@@ -3,9 +3,9 @@ import {__} from '@wordpress/i18n';
 import {A11yDialog} from 'react-a11y-dialog';
 import A11yDialogInstance from 'a11y-dialog';
 import {GiveIcon} from '@givewp/components';
-import {ListTable} from '../ListTable/ListTable';
-import Pagination from '../Pagination/Pagination';
-import {Filter, getInitialFilterState} from '../Filters/Filters';
+import {ListTable} from '../ListTable';
+import Pagination from '../Pagination';
+import {Filter, getInitialFilterState} from '../Filters';
 import useDebounce from '../hooks/useDebounce';
 import {useResetPage} from '../hooks/useResetPage';
 import ListTableApi from '../api';
