+++ src/Views/Components/ListTable/ListTableRows/index.tsx
@@ -2,7 +2,7 @@ import styles from './ListTableRows.module.scss';
 import {__} from '@wordpress/i18n';
 import cx from 'classnames';
 import {useEffect, useState} from 'react';
-import TableCell from '../TableCell/TableCell';
+import TableCell from '../TableCell';
 import {BulkActionCheckbox} from '@givewp/components/ListTable/BulkActions/BulkActionCheckbox';
 import InterweaveSSR from '@givewp/components/ListTable/InterweaveSSR';
