--- src/Views/Components/ListTable/images/RecurringIcon.tsx
@@ -1,20 +1,0 @@
-export default function RecurringIcon({classname = '', ...rest}) {
-    return (
-        <svg
-            role="img"
-            className={classname}
-            viewBox="0 0 23 17"
-            fill="none"
-            xmlns="http://www.w3.org/2000/svg"
-            {...rest}
-        >
-            <path
-                role="presentation"
-                fillRule="evenodd"
-                clipRule="evenodd"
-                d="M0.503387 8.77397L3.50246 8.69925C3.39237 4.28062 6.8821 0.612565 11.3007 0.502479C12.8702 0.463377 14.3412 0.886871 15.5903 1.63599L14.1671 3.1319C13.3262 2.70272 12.3602 2.4767 11.3505 2.50186C8.04157 2.5843 5.4194 5.34046 5.50184 8.64943L8.50091 8.57472L4.60177 12.6731L0.503387 8.77397ZM11.6494 14.4981C14.9584 14.4157 17.5806 11.6595 17.4981 8.35056L14.499 8.42528L18.3982 4.3269L22.4966 8.22603L19.4975 8.30075C19.6076 12.7194 16.1179 16.3874 11.6992 16.4975C10.1297 16.5366 8.65871 16.1131 7.40966 15.364L8.83285 13.8681C9.6738 14.2973 10.6397 14.5233 11.6494 14.4981Z"
-                fill="white"
-            />
-        </svg>
-    );
-}
