--- src/Views/Components/ListTable/images/RenewalIcon.tsx
@@ -1,18 +1,0 @@
-export default function RenewalIcon({className = '', ...rest}) {
-    return (
-        <svg
-            className={className}
-            role="img"
-            viewBox="0 0 19 16"
-            fill="none"
-            xmlns="http://www.w3.org/2000/svg"
-            {...rest}
-        >
-            <path
-                role="presentation"
-                d="M15.2 8H19L14.25 13L9.5 8H13.3C13.3 4.69 10.7445 2 7.6 2C4.4555 2 1.9 4.69 1.9 8C1.9 11.31 4.4555 14 7.6 14V16C3.401 16 0 12.42 0 8C0 3.58 3.401 0 7.6 0C11.799 0 15.2 3.58 15.2 8Z"
-                fill="white"
-            />
-        </svg>
-    );
-}
